+++ R/retrieving.R
@@ -35,9 +35,15 @@
 #' @param matname_colname,matval_colname Names used for matrix names and matrix values.
 #'                                       Defaults are
 #'                                       `PFUPipelineTools::mat_meta_cols$matname` and
-#'                                       `PFUPipelineTools::mat_meta_cols$matval`, respectively.
-#' @param valid_from_version_colname,valid_to_version_colname Names for columns containing version information.
-#'                                                            Defaults are
+#'                                       `PFUPipelineTools::mat_meta_cols$matval`,
+#'                                       respectively.
+#' @param valid_from_version_colname,valid_to_version_colname Names
+#'              for columns containing version information.
+#'              Defaults are
+#'              `PFUPipelineTools::version_cols$valid_from_version`
+#'              and
+#'              `PFUPipelineTools::version_cols$valid_to_version`,
+#'              respectively.
 #' @param conn The database connection.
 #' @param schema The database schema (a `dm` object).
 #'               Default calls `schema_from_conn()`, but
@@ -104,7 +110,6 @@ pl_collect_from_hash <- function(hashed_table,
     # Need to encode foreign keys, because the table in the database has
     # encoded foreign keys
     encode_fks(db_table_name = table_name,
-               conn = conn,
                schema = schema,
                fk_parent_tables = fk_parent_tables)
 
@@ -116,17 +121,15 @@ pl_collect_from_hash <- function(hashed_table,
 
   # Filter on the version, if requested
   if (!is.null(version_string)) {
-    # Figure out the index for the version of interest to us.
-    version_index <- encode_version_string(version_string = version_string,
-                                           version_colname = valid_from_version_colname,
-                                           table_name = table_name,
-                                           conn = conn,
-                                           schema = schema,
-                                           fk_parent_tables = fk_parent_tables)
-    # Filter the outgoing data frame according to the version_index
     out <- out |>
-      dplyr::filter(.data[[valid_from_version_colname]] <= version_index) |>
-      dplyr::filter(.data[[valid_to_version_colname]] >= version_index)
+      filter_on_version_string(version_string = version_string,
+                               db_table_name = table_name,
+                               collect = FALSE,
+                               conn = conn,
+                               schema = schema,
+                               fk_parent_tables = fk_parent_tables,
+                               valid_from_version_colname = valid_from_version_colname,
+                               valid_to_version_colname = valid_to_version_colname)
   }
 
   out <- out |>
@@ -217,10 +220,13 @@ pl_collect_from_hash <- function(hashed_table,
 #' [decode_matsindf()] is called on the downloaded data frame.
 #'
 #' @param db_table_name The string name of the database table to be filtered.
+#' @param version_string A string of length 1 that indicates the desired version.
 #' @param datasets A vector of dataset strings to be retained in the output.
 #'                 `NULL` (the default) returns all datasets.
 #'                 Another interesting option is
 #'                 `PFUPipelineTools::dataset_info$clpfu_iea`.
+#' @param versions A string vector indicating the versions to be downloaded.
+#'                 `NULL`, the default, means to download all versions available in the table.
 #' @param countries A vector of country strings to be retained in the output.
 #'                  `NULL` (the default) returns all countries.
 #'                  Another useful option is `as.character(PFUPipelineTools::canonical_countries)`.
@@ -286,12 +292,21 @@ pl_collect_from_hash <- function(hashed_table,
 #' @param includes_neu_col The name of a column that tells whether non-energy
 #'                         use (NEU) is included.
 #'                         Default is `Recca::psut_cols$includes_neu`.
+#' @param valid_from_version_colname,valid_to_version_colname Names
+#'              for columns containing version information.
+#'              Defaults are
+#'              `PFUPipelineTools::version_cols$valid_from_version`
+#'              and
+#'              `PFUPipelineTools::version_cols$valid_to_version`,
+#'              respectively.
 #'
 #' @return A filtered version of `db_table_name` downloaded from `conn`.
 #'
 #' @export
 pl_filter_collect <- function(db_table_name,
+                              version_string = NULL,
                               datasets = NULL,
+                              versions = NULL,
                               countries = NULL,
                               years = NULL,
                               methods = NULL,
@@ -331,13 +346,30 @@ pl_filter_collect <- function(db_table_name,
                               product_agg_colname = PFUPipelineTools::aggregation_df_cols$product_aggregation,
                               industry_agg_colname =  PFUPipelineTools::aggregation_df_cols$industry_aggregation,
                               chopped_mat_colname = PFUPipelineTools::aggregation_df_cols$chopped_mat,
-                              chopped_var_colname = PFUPipelineTools::aggregation_df_cols$chopped_var) {
+                              chopped_var_colname = PFUPipelineTools::aggregation_df_cols$chopped_var,
+                              valid_from_version_colname = PFUPipelineTools::version_cols$valid_from_version,
+                              valid_to_version_colname = PFUPipelineTools::version_cols$valid_to_version) {
 
   matrix_class <- match.arg(matrix_class)
 
-  out <- dplyr::tbl(src = conn, db_table_name) |>
-    # First, decode the foreign keys with
-    # collect = FALSE to ensure a tbl is returned.
+  out <- dplyr::tbl(src = conn, db_table_name)
+
+  # First, filter according to version string,
+  # if desired.
+  if (!is.null(version_string)) {
+    out <- out |>
+      filter_on_version_string(version_string = version_string,
+                               db_table_name = db_table_name,
+                               conn = conn,
+                               schema = schema,
+                               fk_parent_tables = fk_parent_tables,
+                               valid_from_version_colname = valid_from_version_colname,
+                               valid_to_version_colname = valid_to_version_colname)
+  }
+
+  # Next, decode the foreign keys with
+  # collect = FALSE to ensure a tbl is returned.
+  out <- out |>
     decode_fks(db_table_name = db_table_name,
                schema = schema,
                fk_parent_tables = fk_parent_tables,