+++ R/schema.R
@@ -600,6 +600,7 @@ pl_upsert <- function(.df,
 #' @param db_table_name The string name of the database table where `.df` is to be uploaded.
 #' @param conn An optional database connection.
 #'             Necessary only for the default values of `schema` and `fk_parent_tables`.
+#'             Default is `NULL`.
 #' @param schema The data model (`dm` object) for the database in `conn`.
 #'               See details.
 #' @param fk_parent_tables A named list of all parent tables
@@ -617,7 +618,7 @@ pl_upsert <- function(.df,
 #' @export
 encode_fks <- function(.df,
                        db_table_name,
-                       conn,
+                       conn = NULL,
                        schema = schema_from_conn(conn),
                        fk_parent_tables = get_all_fk_tables(conn = conn, schema = schema),
                        .child_table = PFUPipelineTools::dm_fk_colnames$child_table,
@@ -766,6 +767,8 @@ encode_fks <- function(.df,
 #'                from the database at `conn`.
 #' @param conn An optional database connection.
 #'             Necessary only for the default values of `schema` and `fk_parent_tables`.
+#'             Also necessary if `collect = TRUE`.
+#'             Default is `NULL`.
 #' @param schema The data model (`dm` object) for the database in `conn`.
 #'               See details.
 #' @param fk_parent_tables A named list of all parent tables
@@ -783,8 +786,8 @@ encode_fks <- function(.df,
 #' @export
 decode_fks <- function(.df = NULL,
                        db_table_name,
-                       conn,
                        collect = FALSE,
+                       conn = NULL,
                        schema = schema_from_conn(conn),
                        fk_parent_tables = get_all_fk_tables(conn = conn,
                                                             schema = schema),