+++ R/utilities.R
@@ -297,7 +297,6 @@ inboard_filter_copy <- function(source,
           usual_hash_group_cols = usual_hash_group_cols) |>
     # Decode the foreign keys, so they are human-readable.
     decode_fks(db_table_name = dest,
-               conn = conn,
                schema = schema,
                fk_parent_tables = fk_parent_tables)
 }
@@ -338,6 +337,7 @@ inboard_filter_copy <- function(source,
 #'                      to fk keys.
 #' @param conn A connection to the CL-PFU database.
 #'             Needed only if `fk_parent_tables` is not provided.
+#'             Default is `NULL`.
 #' @param schema The data model (`dm` object) for the database in `conn`.
 #'               Default is `dm_from_con(conn, learn_keys = TRUE)`.
 #'               Needed only if `fk_parent_tables` is not provided.
@@ -364,7 +364,7 @@ inboard_filter_copy <- function(source,
 #'                  fk_parent_tables = fk_parent_tables)
 encode_fk_values <- function(v_val,
                              fk_table_name,
-                             conn,
+                             conn = NULL,
                              schema = schema_from_conn(conn),
                              fk_parent_tables = get_all_fk_tables(conn = conn,
                                                                   schema = schema,
@@ -700,27 +700,29 @@ encode_matsindf <- function(.matsindf,
 #' returning the integer representation of the version string.
 #'
 #' @param version_string The version to be encoded.
-#' @param version_colname A column in which versions are provided.
-#'                        This is likely to be either
-#'                        `PFUPipelineTools::version_cols$valid_from_version`
-#'                        or `PFUPipelineTools::version_cols$valid_to_version`.
-#'                        Either will work.
 #' @param table_name The name of the table in which the `version_string` is found.
-#' @param conn A database connection.
+#' @param conn An optional database connection.
+#'             Necessary only for the default values of `schema` and `fk_parent_tables`.
+#'             Default is `NULL`.
 #' @param schema The schema for the database.
 #'               Default is `schema_from_conn(conn = conn)`.
 #' @param fk_parent_tables The foreign key parent tables.
 #'                         Default is `get_all_fk_tables(conn = conn, schema = schema)`.
+#' @param .version_colname A column in which versions are provided.
+#'                         This column is used internally.
+#'                         Default is `PFUPipelineTools::version_cols$valid_from_version`.
+#'                         `PFUPipelineTools::version_cols$valid_to_version`
+#'                         would also work.
 #'
 #' @export
 #'
 #' @return An integer which is the index for `version_string`.
 encode_version_string <- function(version_string,
-                                  version_colname,
                                   table_name,
-                                  conn,
+                                  conn = NULL,
                                   schema = schema_from_conn(conn = conn),
-                                  fk_parent_tables = get_all_fk_tables(conn = conn, schema = schema)) {
+                                  fk_parent_tables = get_all_fk_tables(conn = conn, schema = schema),
+                                  .version_colname = PFUPipelineTools::version_cols$valid_from_version) {
   # Make sure we have only 1 version
   assertthat::assert_that(length(version_string) == 1,
                           msg = paste("version_string must have length 1",
@@ -728,18 +730,89 @@ encode_version_string <- function(version_string,
   # Make a small data frame that will be used to decode the version.
   mini_df <- tibble::tribble(~c1,
                              version_string) |>
-    magrittr::set_names(version_colname)
+    magrittr::set_names(.version_colname)
   # Encode the versions according to foreign keys
   version_index <- mini_df |>
     encode_fks(db_table_name = table_name,
-               conn = conn,
                schema = schema,
                fk_parent_tables = fk_parent_tables) |>
     # Pull out the only integer we find
-    magrittr::extract2(version_colname)
+    magrittr::extract2(.version_colname)
   # Do some error checking
   assertthat::assert_that(length(version_index) != 0,
                           msg = paste("Didn't find a version that matches", version,
                                       "in PFUPipelineTools::pl_collect_from_hash()."))
   return(version_index)
+}
+
+
+#' Filter a database table based on a version string.
+#'
+#' Filtering a database table based on the version you wish to download
+#' is a common task.
+#' In the CL-PFU database, we store data in a compressed format where
+#' identical data points are not duplicated.
+#' Rather, they are stored in a single row with the `ValidFromVersion`
+#' and `ValidToVersion` columns incremented appropriately.
+#'
+#' The desired version is supplied in the `version_string` argument.
+#'
+#' If both `tbl` and `db_table_name` are provided, `db_table_name` is
+#' ignored.
+#'
+#' @param tbl The `tbl` object that should be filtered.
+#' @param version_string A string of length 1 that indicates the desired version.
+#' @param collect A boolean that tells whether to collect `tbl` from `conn`
+#'                before returning.
+#'                Default is `FALSE`.
+#' @param db_table_name The name of the table to be filtered.
+#' @param conn An optional database connection.
+#'             Necessary only for the default values of `schema` and `fk_parent_tables`.
+#'             Default is `NULL`.
+#' @param schema The database schema (a `dm` object).
+#'               Default calls `schema_from_conn()`, but
+#'               you can supply a pre-computed schema for speed.
+#'               Needed only when `decode_fks = TRUE` (the default).
+#'               If foreign keys are not being decoded,
+#'               setting `NULL` may improve speed.
+#' @param fk_parent_tables Foreign key parent tables to assist decoding
+#'                         foreign keys.
+#'                         Default calls `get_all_fk_tables()`.
+#' @param valid_from_version_colname The name of the ValidFromVersion column.
+#'                                   Default is
+#'                                   `PFUPipelineTools::version_cols$valid_from_version`.
+#' @param valid_to_version_colname The name of the ValidToVersion column.
+#'                                 Default is
+#'                                 `PFUPipelineTools::version_cols$valid_to_version`.
+#'
+#' @returns A filtered version of `tbl`.
+#'
+#' @export
+filter_on_version_string <- function(tbl,
+                                     version_string,
+                                     db_table_name,
+                                     collect = FALSE,
+                                     conn = NULL,
+                                     schema = schema_from_conn(conn = conn),
+                                     fk_parent_tables = get_all_fk_tables(conn = conn, schema = schema),
+                                     valid_from_version_colname = PFUPipelineTools::version_cols$valid_from_version,
+                                     valid_to_version_colname = PFUPipelineTools::version_cols$valid_to_version) {
+
+
+  # Figure out the index for the version of interest to us.
+  version_index <- encode_version_string(version_string = version_string,
+                                         table_name = db_table_name,
+                                         schema = schema,
+                                         fk_parent_tables = fk_parent_tables)
+  # Filter the outgoing data frame according to the version_index
+  out <- tbl |>
+    dplyr::filter(.data[[valid_from_version_colname]] <= version_index) |>
+    dplyr::filter(.data[[valid_to_version_colname]] >= version_index)
+
+  if (collect) {
+    out <- out |>
+      dplyr::collect()
+  }
+
+  return(out)
 }